util: remove Python timeline utilities (#5102)
In general, these have all been replaced by more up to date TypeScript
utilities. In particular, being able to test the timeline using
`timeline.ts` directly vs reimplementing parsing of timelines in Python
is a really important step. It'll make it possible to move to using
network log lines in timeline files and adding new timeline keywords.

The only slightly odd one here is `timeline_aggregator.py`, which does
not have a replacement. However, it's a holdout of trying to parse
timeline files in Python and needs to be rewritten. If timeline formats
change (to add loops, network syncs, etc) then this will break. Removing
this for now and we can resurrect if anybody needs it.