raidboss: Sophia EX changes (#5181)
This isn't as extensive as I expected. The logic contained here is still
basically sound, and there isn't a whole lot of reason to change it
over.

First of all, I changed over the Duplicate/Execute shot calls to be the
full-length versions of the directions. (North as opposed to N,
Southeast as opposed to SE.) There's no reason to keep these shortened
in modern use, now that we have easy indirection available for output
text strings.

Second, I changed over the logic for Execute positioning. The previous
setup was fine, but it would call only one spot if it was Thunder 2.
(Thunder 2 is always cast on exactly one clone during Duplicate.) This
new setup does mean we won't have callouts for safe north/south, but it
guarantees that any safe corners are indicated.

I haven't touched the tilt call logic since there's not much to improve
there. The setup is kind of wonky, but there's no way to call the
tethers differently. Shifting over to calling based on positional data
that now exists in the `14` lines is possible, but it wouldn't replace
the tether logic, since the tethers go out well before the casts do.

Finally, I expanded out the match groups in the timeline so we can
find/replace them more easily, either manually or via script. It is
technically possible to script around the compact matches, but it's
easier just to do this.

Overall, I think we should be fine just changing what's here and then
reacting to anything that's needed.