raidboss: Preliminary Sophia Unreal timeline/triggers (#5194)
This is based on modifications to `sync_files` to parallel what we're
doing with Criterion dungeons. Overall it seems to have been successful.
I was able to get a small amount of testing in last night before
maintenance, and most stuff worked.

The portion of the timeline between the Scales of Wisdom cast and the
first set of post-intermission clones is messy. I'll need to do more
research to find the appropriate places to sync that. Regardless, after
the first set of clones, everything seems to sync nicely.