raidboss: p3 add break tethers for even tethers
This adds a break tether callout for when the magic vulnerability up from the Patch spell expires.

Potential for this to be expanded to call the tower to run to with merging of my other PR.

I don't really see it necessary to call the first break here since it happens naturally. For the last set I am not too sure on, I only know that the regression that breaks there changes order. Not sure how these triggers impact P5, might need to condition these to only run in P3?