oopsy: remove unused AbilityIdMap
This used to be used at one point, but hasn't been used for a good
while.  Oopsy has grown a lot more complicated since this originally
existed and the ability id map would need to be hooked up in
quite a few places (including death reports and oopsy triggers) in
order for this feature to be useful again.