raidboss: TOP P5 delta tether + beyond defense filter (#5303)
The effects are unique for delta, so there wasn't a need to use phase
filter.

I think maybe in the future it would say something more for the blue
tether players like blue tether with some description about the world
they may be with. For certain strats, players are using near world or
distant world as a method to say which blue tether pair goes where.