raidboss: TOP add config option + rotation to omega dodge
Because the rotation direction depends on which side you
have dodged to, it's impossible to write this PR without
having the user pick which dodge they are doing.

This promotes the dodge direction to a new option (defaulting
to far, since that seems to be the PF approach) and adds
in the rotation direction as well.

Because there's an option, there's no longer any combo
output strings for staff sword, and the correct one is picked.