raidboss: TOP P6 Add Cosmo Arrow Exa Calls (#5377)
There are NPCs that are added at the first Cosmo Arrow that then cast
7BA3 Cosmo Arrow. The position that these NPCs spawn in at or where they
start casting at can be used to determine if it is in or out exasquare
first.

There are 4 NPCs used for the outer exasquare and two NPCs for the inner
exasquare.

The second Cosmo Arrow will use these same NPCs.

This calls initial in/out before any visible aoes. The rest are timed
with the ability.