raidboss: TOP Unlimited Wave Cannon Dodge + Clock (#5370)
I noticed I could use some code I wrote for DSR gigaflares here that
should also work for determining the exaflare clock here.

This calls the card/intercard that is safe and next to the first
exaflare. In the same output, it calls the direction
(clock/counterclock) the exaflares spawned in which is where the player
will need to run later.