raidemulator: Refactor `CombatantState` object (#5387)
This PR is in preparation for `combatantMemory` lines.

Because the state tracker will now need to track all possible
`PluginCombatantState` properties, it no longer makes sense to track a
limited set for most other types of lines and a full state for those
specific lines.

As such, the custom properties on `CombatantState` have been removed in
favor of just implementing the `PluginCombatantState` interface with the
carried over addition of a `targetable` property to maintain
functionality.

Additionally, logic for constructing and cloning states was greatly
simplified to use `Object.assign` instead.

Some duplicate properties were removed from `Combatant` in favor of the
tracked state entries in `CombatantState`, and associated locations were
updated to instead read the initial combatant state's version of those
properties. This should result in more correct behavior when
`getCombatants` is used, because name changes are tracked properly now
for example.

The vast majority of the added lines are in a new
`resources/world_id.ts` file, required to translate `ID` <=> `Name`.
Excluding that file, the initial commit for this PR contains 166
additions and 205 deletions.

This PR needs more testing before being merged.

- I have only tested with US DC names, for example, so I'm not sure if
world name lookup will work as expected for `CN`/`KO`/`JP` servers.
- I have not tested for logs that may have weird values for some entries
that would potentially result in `parseFloat` returning `NaN`.
- I have only tested recent content (DSRU, P8S, and TOP). Some older
content should probably be tested.