util: support jumps in test_timeline (#5435)
This avoids having test_timeline report events as being missed when
intentionally jumping over them.

I'm not 100% sure about this implementation, my understanding of this
code is still pretty surface level, but it seems to work for me on a few
logs I tested it against (mostly my new p9n stuff).

Also not sure if I should do `>=` or `>`. The latter still reports a
missed entry for the line you're jumping to, e.g. this:

```
100.0 "foo" jump 200.0
150.0 "bar"
200.0 "foo"
201.0 "baz"
```

would output something like this with `>`:

```
       0.000 |    1 | 100.0 "foo" jump 200.0
      Missed |    3 | 200.0 "foo"
       0.000 |    4 | 201.0 "baz"
```

But with `>=` it outputs:

```
       0.000 |    1 | 100.0 "foo" jump 200.0
       0.000 |    4 | 201.0 "baz"
```

I'm not sure exactly what the behaviour should be like if you have
multiple entries at the same time that you're jumping to, like:

```
100.0 "foo" jump 200.0
150.0 "bar"
200.0 "foo"
200.0 "baz"
```