raidboss/oopsy: add P9N (#5432)
This is my first attempt at a timeline, so there's probably some issues
here.

I've run this with `test_timeline.ts` against 17 different runs that I
did and didn't see anything particularly concerning, there's a few
abilities that are sometimes about 0.2 seconds off, but for the majority
of the logs it's more or less spot on.

I might've gone a little bit overboard with the forking/merging of
timelines in the timeline, since the actual casts are about the same
either side of the fork, but I decided to leave it in for the first
review since I already did the work of splitting it up. Happy to merge
it back together, though, if that's preferred, the timeline of this
fight is actually quite linear if you ignore the fire vs ice parts.

I did a test run as well with this, and didn't notice anything obviously
wrong, though I don't have a good feel for what's good when it comes to
triggers and what to include/not include (especially for things like
Burst where you get "Move Away" immediately followed by "Stack on X").

I also included a full table of all the action IDs I discovered when
going through this. Not sure if you want that or not, but I wrote it for
myself in order to make sense of the timeline, so I figured I'd include
it in case it was useful.