raidboss: add gameLogType field for GameLog triggers
This is part of #4986.

This is an alternative PR to #5451.

In order to switch entirely over to `netRegex: { params }` style
triggers, this PR adds a `gameLogType` field for `GameLog` triggers
only.  If specified, it will add the right `code`.

Downsides: kind of awkward if `code` is already specified (and this
will throw an error)

Other options:
* add a bunch more types so you could say `type: 'GameLogEcho'`,
  but this seemed much more hairy than was worth and the TriggerTypes
  type is already very very long
* export the codes from some common location, import them everywhere,
  but this is even more nonsense to import so that user triggers
  can eval properly