raidboss: P12S Timeline (#5518)
WIP-ish timeline, not really sure of the procedure for these sorts of
things. Doorboss has been extensively tested in-game and is working just
fine, but Pallas is untested ingame, as it was created mainly from
watching vods and ACT logs.

However, both Pallas and Doorboss are running just fine when tested
against multiple FFLogs timelines using the test_timeline util.