raidboss: p10s fixups after #5562
In attempting to clean up messages that don't make sense,
e.g. "Partners => Role Stack" I removed the condition check
that prevented it from being double called out.

Also, move the counter to the reset location.  This is not a
logic change, but it makes sense to have the counter on the cast.