raidboss: p10s Jade Passage trigger box/line
This uses the AddCombatant log lines instead of cast since it comes sooner.


Just based on the y values and not having tested in game, largest y value I assume is the lines, so using a map to quickly determine from one of the Arcane Spheres where it is located.

Boxes:
88.00
92.00
96.00
100.00
104.00
108.00
112.00

Lines:
86.00
90.00
94.00
98.00
102.00
106.00
110.00
114.00

This is also assuming the data in the log line is accurate, otherwise OverlayPlugin would be needed. It seems accurate to me and they get spawned in and removed between parts so no need to check for updated location.