raidboss: add support for P12S Engravements (#5587)
Not sure if I did something dumb while rebasing or Visual Studio did,
but it was a messy unwind, and doing a new PR was POLR. So I'm killing
the old PR (#5580 ) and opening a new one. This includes all revisions
incorporating feedback from the initial code review. Apologies for any
added work.