raidboss: add P12s Engravement 1 Tower trigger (#5597)
Add P12s Engravement 1 Tower triggers. 
> This trigger is based on part of the functionality of #5587 PR, and
later merges may involve conflict merges, which I will resolve at that
time.

In Engravement 1 (Paradeigma 2), 2 players receive lightTower and 2
players receive darkTower, 2 players need to guide the light beam and 2
players need to guide the dark beam.

The operator of the beam extends the beam directly from the outside. The
beam is attenuated until the jagged line disappears.

The people in the tower find the people who have the opposite attribute
to the debuff and put them in four places. At NE NW SE SW as a # shape.
The position of outside Anthropos is fixed by two situation. `{[97, 75],
[125, 97], [103, 125], [75, 103]} and {[103, 75], [125, 103], [97, 125],
[75, 97]}`. The Anthropos will cast 'Searing Radiance' for light beam
and 'Shadowsear' for dark beam. We use those as a trigger for Tower
players place the Tower.

When debuffs expire and towers drop, their debuff changes to lightTilt
or darkTilt (same as tower color).
At the same time the towers drop, the 4 tethered players receive
lightTilt or darkTilt depending on their tether color.


![example](https://github.com/quisquous/cactbot/assets/12003087/6d2eb614-a04a-4225-97be-5968cd3279fa)

---------

Signed-off-by: 柚木鉉 <GloomyGhost@foxmail.com>
Co-authored-by: Adrienne Walker <enne@quisquo.us>