raidboss: improvements to p12sp2 classical concepts (#5632)
Tested extensively against logs & vod, but could use some additional
testing in game and validation on the timings.

Right now, for classical2, it will display the initial position for 7
seconds as an infoText, and right about as that expires, it will then
display the inverted (actual) position as an alertText. I left those as
static values, but I'm undecided whether it would make more sense to add
a config option to allow the user to specify when the alert should
'transition' from initial to actual.

Admittedly, I made a stylistic choice to do a single output trigger that
covers the output for classical1 as well as the initial and actual calls
for classical2. It adds some trivial complexity around delay/duration,
but the primary benefit I saw in this approach was a single set of
customizable output strings, rather than multiple triggers users would
separately need to modify. (It also removes the need for some
duplicative code.) But I understand reasonable minds may differ on that
choice.