util/test: add missing netRegex params to find translations (#5667)
This adds extra functionality to the existing netRegex param translation
to indicate missing translations.

This allows test_trigger and find_missing_timeline_translations to both
re-use the actual path at runtime that translates. This removes bespoke
logic in test_trigger to do this same work and adds this logic to
find_missing_timeline_translations to report missing netRegex
translations (that it previously did not do).

Previously also, doing the translation of netRegex params overwrote the
params in place with the new translation. This wasn't technically an
issue because of the post-collision checks, which mean that you can
apply translations any number of times and still get the same result as
applying it once. However, this is very confusing and subtle behavior
and so now it makes a copy of the params instead.