raidboss: add `forcejump` timeline command (#5577)
Rather than adding a `loop`, add a variant of `jump` called `forcejump`
to the timeline language. If the sync does not occur, the jump will
always happen when the time passes. This allows the timeline controller
to automatically add loop lookahead rather than manually specifing it.

There's some manual testing of this in the Frankenstein manual
`test.txt` test case.

To track this, `nextEvent` now has a whole bunch more state to track
where it is in the lookahead so that when adding additional events as
timer bars, it doesn't duplicate its efforts.

Additionally, since the same event could be added twice (or more) times
on the screen, they need to have different ids (now strings instead of
numbers, that include the "jump count" of the number of times we've gone
through a lookahead jump to differentiate them from their pre-jump
selves). Similarly, they also need different sort keys, although this is
kind of a hack for now.

This also adds some grumpy documentation of a bunch of issues found in
the code and potential followups.
Sorry that this code is a mess, but I didn't want to rewrite the world
at this time.

Closes #5144.