util: fix async bugs with find missing translations
The "code" missing translations finder wasn't properly async
and so would continue to accumulate missing translations as
the rest of the coverage report continued.

The find missing translations script (standalone) handles
this by not ending automatically, but the coverage report
would (for example) list 4 entries but then say 21 errors
as the the counting of the errors came after the entries
were written.