raidboss/oopsy: Thordan EX (#5819)
(This will close #5742 )

This should largely be self-explanatory. We are fortunate that this
content is old enough that positions and headings are accurate without
having to do any work for them.

I've tested this live both solo and with others (thanks quisquous!) but
it will definitely need some polish here and there. In particular, some
complex mechanics feel very noisy. It would be nice to figure out what
alerts can be downgraded to info.

Note that currently there's no support for keeping Defamation and Pierce
players out of the stack during the meteor trio. This can be done
without too much effort but I have no brain left for today.