raidboss/oopsy: Add Singularity Reactor Unreal (#5831)
Updated sync_files with data sourced from FFLogs. The skill IDs line up,
but some of the actors appear to have changed. Notably, the Comet and
Heavy Impact abilities are now sourced from Thordan rather than various
knights, and The Light Of Ascalon is also sourced by Thordan. I don't
know whether this is an actual change to the encounter or an artifact of
FFLogs/ACT needing updates.

This is fully untested in-game, and I won't have time until Friday, but
here it is.