raidboss: update wording for ex7 miasmic (#5847)
Most players seem to dodge north for this (even though it makes the
spread tough). Also, the current version orders the outputs by their
ordering in the array, which means that sometimes the "north" direction
is first and sometimes second.

To add clarity, only consider safe spots on the north side of the room,
and then add explicit wording for where to go.

This is maybe controversial as "east wall" could be ambiguous between
ENE and ESE but it felt like to me that if most people dodge north, we
can just say the "north" safe spot with details in parentheses. I found
this to be more clear, at least.

This output strings change would also let people change to using marker
names in the output strings more easily.