util: improve coverage report for Chinese/Korean
Currently the "missing translations" lists hundreds of errors
for each timeline entry that can't be translated because this
content doesn't exist yet.

To make it easier to find issues with content that can be fixed,
create a new category of error for missing 'replaceSection'
and separate out those errors into their own column on the sheet.

This also makes it possible to show better "how many fights are
translated" numbers for other languages by not including fights
that don't exist yet in that region.