oopsy: update oopsy code to use player objects (#5906)
Followup to #5861.

There's no "output strings" in oopsy, but it does use ShortName in a few
places. Move this over to data.party.member(x).toString() in order to
potentially support an oopsy option for how to display names.

This also changes some logic for mit tracking and raise tracking to use
full names instead of short names when storing and comparing, only using
the shortened version for output.

This also creates the party tracker earlier so that it can be passed to
all the things that want to use it, and so all of oopsy can indirect
through `PartyTracker` to find out short names.