lint: add errors for using deprecated functions (#5915)
This is mostly to prevent the use of data.ShortName (when triggers
should be using data.party.member instead), but apparently there were a
lot of deprecated function uses to clean up while I was here. Thanks,
lint.

Apparently also, Regexes.addedCombatantFull was not the same as
Regexes.addedCombatant, even though the NetRegexes versions are the
same. Fixed that here.