util: Add basic network log seal support in timeline tools (#5947)
This wasn't as simple as I had hoped, but it does seem to work as it is.
There's still a bit of fragility in `make_timeline` with this
implementation, since it still relies on the seal name being present in
game log form. We could change this to make "name from game log"
optional in the `assembleTimelineStrings()` function, but that would
involve some complex conditions to allow for use of either the network
seal or the game log seal.