raidboss/test: improve testing of timeline netregex
This is a follow-up to #5962, which itself is related to #5939.

I forgot that `find_missing_timeline_translations.ts` is only half
of the testing picture and `test_timeline.ts` needed to be updated
too. This was found by trying to test more lines and having test
timeline complain that InCombat regex lines were not translated.